feat(documents): migrate documents page to server/client architecture
- Convert documents page from client-only to hybrid server/client components
- Add server component for static header and metadata (SEO optimization)
- Extract dynamic functionality to DocumentsContent client component
- Create DocumentsSkeleton for loading states with Suspense streaming
- Add /api/documents route with GET endpoint for documents and applications
- Implement cross-page data dependency (documents require applications)
- Maintain all existing functionality: file upload, search, document grid, statistics
- Preserve localStorage as fallback during transition period
- Follow established Next.js 14 App Router patterns for performance
